fix SearchGithubRepositoryViewController
