--- iOSEngineerCodeCheck/Scene/SearchGithubRepository/SearchGithubRepositoryViewController.swift
@@ -1,88 +1,0 @@
-//
-//  SearchGithubRepositoryViewController.swift
-//  iOSEngineerCodeCheck
-//
-//  Created by 史 翔新 on 2020/04/20.
-//  Copyright © 2020 YUMEMI Inc. All rights reserved.
-//
-
-import UIKit
-
-class SearchGithubRepositoryViewController: UITableViewController, UISearchBarDelegate {
-
-    @IBOutlet weak var searchBar: UISearchBar!
-
-    var repositories: [[String: Any]] = []
-    var task: URLSessionTask?
-    var word: String!
-    var url: String!
-    var idx: Int!
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-        searchBar.text = "GitHubのリポジトリを検索できるよー"
-        searchBar.delegate = self
-    }
-
-    func searchBarShouldBeginEditing(_ searchBar: UISearchBar) -> Bool {
-        // ↓こうすれば初期のテキストを消せる
-        searchBar.text = ""
-        return true
-    }
-
-    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-        task?.cancel()
-    }
-
-    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-        word = searchBar.text!
-        if word.count != 0 {
-            url = "https://api.github.com/search/repositories?q=\(word!)"
-            task = URLSession.shared.dataTask(with: URL(string: url)!) { (data, _, _) in
-                if let obj = try? JSONSerialization.jsonObject(with: data!) as? [String: Any] {
-                    if let items = obj["items"] as? [[String: Any]] {
-                    self.repositories = items
-                        DispatchQueue.main.async {
-                            self.tableView.reloadData()
-                        }
-                    }
-                }
-            }
-        // これ呼ばなきゃリストが更新されません
-        task?.resume()
-        }
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return repositories.count
-    }
-
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = UITableViewCell()
-        let repository = repositories[indexPath.row]
-        cell.textLabel?.text = repository["full_name"] as? String ?? ""
-        cell.detailTextLabel?.text = repository["language"] as? String ?? ""
-        cell.tag = indexPath.row
-        return cell
-    }
-
-    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        // 画面遷移時に呼ばれる
-        idx = indexPath.row
-        transitionToGithubRepository()
-    }
-
-    // TODO: Routerに移動させる
-    /// Githubリポジトリの詳細に遷移
-    private func transitionToGithubRepository() {
-        guard let githubRepositoryVC = UIStoryboard(
-            name: "GithubRepository",
-            bundle: nil
-        ).instantiateInitialViewController() as? GithubRepositoryViewController else {
-            fatalError("Not found GithubRepositoryViewController")
-        }
-        githubRepositoryVC.vc1 = self
-        navigationController?.pushViewController(githubRepositoryVC, animated: true)
-    }
-}
