+++ iOSEngineerCodeCheck/Scenes/SearchGithubRepository/SearchGithubRepositoryViewController.swift
@@ -1,0 +1,72 @@
+//
+//  SearchGithubRepositoryViewController.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by 史 翔新 on 2020/04/20.
+//  Copyright © 2020 YUMEMI Inc. All rights reserved.
+//
+
+import UIKit
+
+final class SearchGithubRepositoryViewController: UITableViewController {
+    private var presenter: SearchGithubRepositoryPresenterInput?
+    @IBOutlet private weak var searchBar: UISearchBar!
+
+    var repositories: [[String: Any]] = []
+    var idx: Int!
+
+    func inject(presenter: SearchGithubRepositoryPresenterInput) {
+        self.presenter = presenter
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+        searchBar.text = "GitHubのリポジトリを検索できるよー"
+        searchBar.delegate = self
+    }
+
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return repositories.count
+    }
+
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = UITableViewCell()
+        let repository = repositories[indexPath.row]
+        cell.textLabel?.text = repository["full_name"] as? String ?? ""
+        cell.detailTextLabel?.text = repository["language"] as? String ?? ""
+        cell.tag = indexPath.row
+        return cell
+    }
+
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        // 画面遷移時に呼ばれる
+        idx = indexPath.row
+        presenter?.didSelectRepository(row: indexPath.row)
+    }
+}
+
+// MARK: - UISearchBarDelegate
+extension SearchGithubRepositoryViewController: UISearchBarDelegate {
+    func searchBarShouldBeginEditing(_ searchBar: UISearchBar) -> Bool {
+        // ↓こうすれば初期のテキストを消せる
+        searchBar.text = ""
+        return true
+    }
+
+    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
+        presenter?.searchTextDidChange()
+    }
+
+    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
+        presenter?.searchGithubRepositries(word: searchBar.text)
+    }
+}
+
+// MARK: - SearchGithubRepositoryPresenterOutput
+extension SearchGithubRepositoryViewController: SearchGithubRepositoryPresenterOutput {
+    func didSearchGithubRepositories(_ repositories: [[String: Any]]) {
+        self.repositories = repositories
+        tableView.reloadData()
+    }
+}
